pyproject.toml
[project]
name = "horizons"
version = "0.1.0"
description = "Automated follower insight tool"
readme = "README.md"
authors = [{ name = "HalFTeen" }]
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "requests>=2.31.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "trafilatura>=1.7.0",
    "sqlite-utils>=3.37",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.3",
    "rich>=13.7.0",
    "yt-dlp>=2024.1.1",
    "faster-whisper>=1.0.0"
]

[project.scripts]
horizons = "horizons.cli:app"

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
